+++ pages/view-departments.php
@@ -1,5 +1,5 @@
 <?php
-
+include("../auth/security.php");
 include("../config/database.php");
 
 $sql = "SELECT * FROM departments";
@@ -82,17 +82,30 @@ $result = mysqli_query($connection, $sql);
                                 </td>
                                 <td>
 
-                                    <a
-                                        href="edit-departments.php?id=<?php echo $department['department_id']; ?>"
-                                        class="btn btn-primary btn-sm">
-                                        Edit
-                                    </a>
+                                    <?php if (isAdminOrSubAdmin()) { ?>
 
-                                    <a
-                                        href="delete-departments.php?id=<?php echo $department['department_id']; ?>"
-                                        class="btn btn-danger btn-sm">
-                                        Delete
-                                    </a>
+                                        <a
+                                            href="edit-departments.php?id=<?php echo $department['department_id']; ?>"
+                                            class="btn btn-primary btn-sm">
+
+                                            Edit
+
+                                        </a>
+
+                                    <?php } ?>
+
+                                    <?php if (isAdmin()) { ?>
+
+                                        <a
+                                            href="delete-departments.php?id=<?php echo $department['department_id']; ?>"
+                                            class="btn btn-danger btn-sm"
+                                            onclick="return confirm('Are you sure?');">
+
+                                            Delete
+
+                                        </a>
+
+                                    <?php } ?>
 
                                 </td>
