+++ pages/view-doctors.php
@@ -1,6 +1,9 @@
 <?php
 
+
+include("../auth/security.php");
 include("../config/database.php");
+
 
 $sql = "SELECT
             doctors.doctor_id,
@@ -95,17 +98,30 @@ $result = mysqli_query($connection, $sql);
                                 </td>
                                 <td>
 
-                                    <a
-                                        href="edit-doctors.php?id=<?php echo $doctor['doctor_id']; ?>"
-                                        class="btn btn-primary btn-sm">
-                                        Edit
-                                    </a>
+                                    <?php if (isAdminOrSubAdmin()) { ?>
 
-                                    <a
-                                        href="delete-doctors.php?id=<?php echo $doctor['doctor_id']; ?>"
-                                        class="btn btn-danger btn-sm">
-                                        Delete
-                                    </a>
+                                        <a
+                                            href="edit-doctors.php?id=<?php echo $doctor['doctor_id']; ?>"
+                                            class="btn btn-primary btn-sm">
+
+                                            Edit
+
+                                        </a>
+
+                                    <?php } ?>
+
+                                    <?php if (isAdmin()) { ?>
+
+                                        <a
+                                            href="delete-doctors.php?id=<?php echo $doctor['doctor_id']; ?>"
+                                            class="btn btn-danger btn-sm"
+                                            onclick="return confirm('Are you sure?');">
+
+                                            Delete
+
+                                        </a>
+
+                                    <?php } ?>
 
                                 </td>
